Make events extensible
