+++ frontend/src/metabase-types/analytics/embed-flow.ts
@@ -1,5 +1,3 @@
-import type { EmbedResourceDownloadOptions } from "metabase/public/lib/types";
-
 type EmbedFlowParams = {
   locked?: number;
   enabled?: number;
@@ -13,7 +11,7 @@ type EmbedFlowAppearance = {
   theme?: string;
   font?: string;
   downloads?: boolean | null;
-  enabled_download_types?: EmbedResourceDownloadOptions | null;
+  enabled_download_types?: { pdf?: boolean; results?: boolean } | null;
 };
 
 type EmbedFlowEventSchema = {
