+++ frontend/src/metabase/home/components/Onboarding/analytics.ts
@@ -1,6 +1,25 @@
 import { trackSimpleEvent } from "metabase/lib/analytics";
+import type { ValidateEvent } from "metabase-types/analytics/event";
 
 import type { ChecklistItemCTA, ChecklistItemValue } from "./types";
+
+type OnboardingChecklistItemExpandedEvent = ValidateEvent<{
+  event: "onboarding_checklist_item_expanded";
+  triggered_from: ChecklistItemValue;
+}>;
+
+type OnboardingChecklistItemCTAClickedEvent = ValidateEvent<{
+  event: "onboarding_checklist_cta_clicked";
+  triggered_from: ChecklistItemValue;
+  event_detail: ChecklistItemCTA;
+}>;
+
+declare module "metabase-types/analytics/event" {
+  interface SimpleEventExtensions {
+    onboardingChecklistItemExpanded: OnboardingChecklistItemExpandedEvent;
+    onboardingChecklistItemCTAClicked: OnboardingChecklistItemCTAClickedEvent;
+  }
+}
 
 export const trackChecklistItemExpanded = (value: ChecklistItemValue) => {
   trackSimpleEvent({
